Refactored video upload and embed extension code
