--- src/components/Editor/CustomExtensions/UnifiedVideo/constants.js
@@ -1,11 +1,0 @@
-export const DEFAULT_VIDEO_WIDTH = 502;
-export const DEFAULT_VIDEO_HEIGHT = "fit-content";
-export const DEFAULT_EMBED_WIDTH = 500;
-export const DEFAULT_EMBED_HEIGHT = 281;
-export const DEFAULT_ALIGN = "left";
-export const DEFAULT_FRAME_SETTINGS = {
-  frameBorder: "0",
-  allow:
-    "accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture",
-  allowfullscreen: "allowfullscreen",
-};
